Implement treegrid to display parent-child-relationship in project list (#328)
* Implement treegrid for parent-child-relationship

Removed the hierarchical view with subtables in favor of the default
table with an added treegrid to display the hierarchy.
Also includes lazy-loading sub-levels.

Signed-off-by: RBickert <rbt@mm-software.com>

* Implement dummy children for better lazy-loading

Dummy children are now used instead of loading a level ahead to reduce
API calls and optimize the loading times of a projects next level.

Pagination not working correctly because dummy children need to be
removed and removing from a bootstrap-table decrements the total row
count.

Signed-off-by: RBickert <rbt@mm-software.com>

* Remove dummy children for lazy-loading

Removes dummy children for lazy-loading to avoid the need of removing
data from the table, which caused problems with the pagination.

Now renders the expander button for a project, if it has children which
meet the current filter criteria (e.g. inactive, only applications)

Signed-off-by: RBickert <rbt@mm-software.com>

* Retrieve available parents only when needed

Only fetches the available parents for the `ProjectDetailsModal` and the
`ProjectCreateProjectModal` when the modals are opened, instead of at
the startup of the page, to reduce unnecessary loading of data.

Signed-off-by: RBickert <rbt@mm-software.com>

* Fix checking visibility of wrong column

Signed-off-by: RBickert <rbt@mm-software.com>

* Implement flat view and search

Project list is a hierarchical view per default, can now be switched to
flat view.

Automatically switches to flat view while searching. Disables switching
views while searching.

Signed-off-by: RBickert <rbt@mm-software.com>

* Toggle `FlatView` switch to `on` when searching

Toggles the state of the `showFlatView` switch to `on` when searching
and saves the state before searching to switch back to it after
searching.

Signed-off-by: RBickert <rbt@mm-software.com>

* Switch to `uuid` and `parentUuid` for hierarchy

Instead of using a projects `id` and `pid` to map the relationship
between children and parents, a projects `uuid` and `parentUuid` are now
hashed with `MurmurHash2` and then used to create the treegrid.
(A treegrid's `idField` and `parentIdField` must be numbers and
not strings to correctly work)

Signed-off-by: RBickert <rbt@mm-software.com>

* Remove unused parameter of `onClickRow`

Signed-off-by: RBickert <rbt@mm-software.com>

* Change parent input form

Change parent input form to `vue-multiselect` and the API
endpoint to
`/api/v1/project?searchText=<fieldInput>&excludeInactive=false`

Signed-off-by: RBickert <rbt@mm-software.com>

* Fix API endpoint including inactive projects

Signed-off-by: RBickert <rbt@mm-software.com>

Signed-off-by: RBickert <rbt@mm-software.com>