Show update available icon in dependency graph
`Show update information` can now be toggled in the dependency graph.
This will show a yellow warning icon, if a component's version is not
the same as its latest version from the configured repositories.

(Works the same as in the project's component list, but without showing
the green icon, if the component's version is the same as the latest
version)

Signed-off-by: RBickert <rbt@mm-software.com>