Global Audit Views: Policy Violations
Adds a new view in the sidebar, Policy Violation Audit, which shows
every policy violation filtered by ACLs and other optional filters.
This view makes it easier to get an overview of all policy violations in
all the projects with the possibility to filter them and only show
relevant violations.

`VIEW_POLICY_VIOLATION` permission is required to access the view

Signed-off-by: RBickert <rbt@mm-software.com>