Make removing parent relationship more convenient
The parent relationship of a project can be removed in the frontend by
selecting the currently selected parent for a second time.
To make this removal more convenient, the ProjectDetailsModal now
preloads the parent selection with the selected parent instead of an
empty list, where you'd have to search for the parent to remove it.

Signed-off-by: RBickert <rbt@mm-software.com>