Fix redundant requests to `/api/v1/component` when loading project page
`onlyOutdated` and `onlyDirect` were initialized with non-boolean values. Both variables have a watch on them, triggering a refresh of the components table upon change. Both variables are used as model for `c-switch` components, causing their value to transition to `false` immediately, triggering the watches and consequently table refreshes.

This behavior could cause requests to be abandoned before their content was fully consumed, resulting in `EofException`s in the API server.

Signed-off-by: nscuro <nscuro@protonmail.com>